+++ src/parse/vcf.rs
@@ -8,7 +8,7 @@ use crate::parse::compounds::parse_compounds;
 use crate::parse::conservations::parse_conservations;
 use crate::parse::coordinates::parse_coordinates;
 use crate::parse::filters::parse_filters;
-use crate::parse::frequencies::{add_frequencies, parse_frequencies};
+use crate::parse::frequencies::{add_frequencies, parse_frequencies, parse_sv_frequencies};
 use crate::parse::fusions::set_fusion_info;
 use crate::parse::genetic_models::parse_genetic_models;
 use crate::parse::genotypes::{parse_genotypes, validate_sample_mapping};
@@ -199,6 +199,14 @@ pub fn process_vcf(
             }
 
             _ => {}
+        }
+
+        if matches!(category, VariantCategory::Sv | VariantCategory::CancerSv) {
+            let sv_frequencies = parse_sv_frequencies(&record);
+
+            if let Ok(frequencies) = variant.get_document_mut("frequencies") {
+                frequencies.extend(sv_frequencies);
+            }
         }
 
         let parsed_transcripts = parse_vep_transcripts(&record, &vep_header, &mut variant);